Fixing bugs in Quizzes page
