+++ src/Components/Error.js
@@ -1,0 +1,8 @@
+export default function Error(props) {
+  return (
+    <div>
+      <h2>I don't think you're meant to get to this page....</h2>
+      <h1>404 Page not found?</h1>
+    </div>
+  );
+}
